update LLM classes
